+++ times.py
@@ -1,0 +1,8 @@
+import datetime
+
+
+month    = datetime.datetime.now().month
+year     = datetime.datetime.now().year
+day      = datetime.datetime.now().day
+hour     = datetime.datetime.now().hour
+minute   = datetime.datetime.now().minute
